refactor: update vendor and medicine entity mappings and data types
- Removed legacy `Product` entity, repository, and service implementation.
- Refactored `Vendor` entity and related DTOs to use `Integer` for IDs instead of `Long` to align with the database schema.
- Renamed table mappings for `Medicine` (to `vendor_medicine`) and `Stock` (to `vendor_medicine_price`).
- Updated `OrderItem` entity to map `medicine_id` column to `vendor_medicine_id`.
- Adjusted `BucketServiceImpl`, `OrderServiceImpl`, and various repositories to handle the updated `vendorId` data types and entity relationships.
- Cleaned up unused imports and legacy SQL comments in entity classes.

+++ src/main/java/com/trego/dao/entity/PreOrder.java
@@ -60,7 +60,7 @@ public class PreOrder {
     private LocalDateTime modifiedAt;
     
     @Column(name = "selected_vendor_id", nullable = true)
-    private Long selectedVendorId;
+    private Integer selectedVendorId;
 
     @PrePersist
     protected void onCreate() {
